Implement user can view service suggested and paid

+++ lib/application_layer/screens/booking/see_booking_detail_payment.dart
@@ -13,7 +13,7 @@ class SeeBookingDetailPayment extends StatefulWidget {
 }
 
 class _SeeBookingDetailPaymentState extends State<SeeBookingDetailPayment> {
-  int _bookingPrice = 0;
+  double _bookingPrice = 0;
 
   _getBookingPrice() async {
     var response = await BookingService().getBookingPrice();
